SASS-5065: Refactored the code
