+++ app/services/GetIncomeSourcesService.scala
@@ -20,9 +20,10 @@ import common.IncomeSources._
 import connectors._
 import models._
 import models.cis.{AllCISDeductions, CISSource}
-import models.employment.{AllEmploymentData, OtherEmploymentIncome}
+import models.employment.AllEmploymentData
 import models.gains.InsurancePoliciesModel
 import models.gifts.GiftAid
+import models.otheremployment.OtherEmploymentIncome
 import models.pensions.Pensions
 import models.statebenefits.AllStateBenefitsData
 import play.api.Logging
